Updated Readme
This data includes temperature metrics, precipitation metrics, wind speed, various weather conditions, holiday indicators, and customer metrics from January 1, 2011, through December 21, 2018. This data set was found on Kaggle.com and created from data provided by Capital Bikeshare, NOAA’s National Climatic Data Center, and DC Department of Human Resources. Insights gained from this data can help make more strategic decisions about the supply and demand for bikes, based on the weather in Washington, D.C., by identifying relationships between the weather and usage, allowing business partners to plan effectively for business needs. These relationships will be used for predictive analysis to forecast the demand needs based on statistically significant weather conditions. As part of the business solution, the users will be provided with meaningful data visualizations to assist in decision-making and a simple python program that will allow the users to input weather information and output a forecast.
R was used to run a statistical analysis to determine the significance of the variables, tableau provided meaningful data visualization that was both easy to read and manipulate by the user, and python was used to create a program allowing a user to determine a prediction of the customer demand based on weather conditions.